requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
fpdf
pandas
matplotlib
seaborn
python-multipart